requirements.txt
#requires python>=3.4
# which ones are for croppers
# which ones for inference
# which ones for training
# which ones for post-processing
pytest
numpy
python-dateutil
google-api-python-client
# googlemaps
oauth2client
requests
ExifRead
Pillow
psycopg2
twilio
psutil
tensorflow==1.15
scikit-image